requirements.txt
coreapi>=2.0.0,<2.1
django>=1.10.1,<1.11
django-cors-middleware>=1.3.0,<1.4
djangorestframework>=3.4.6,<3.5
django-filter>=0.9.2,<1
drf-jwt-knox>=0.1,<0.2

# We need the latest drf-extensions for better nested-route support
git+https://github.com/ssaavedra/drf-extensions.git#egg=drf-extensions

